requirements.txt
# Core - minimum requirements for demo
streamlit==1.28.1
streamlit-folium==0.15.0
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0
folium==0.14.0
branca==0.6.0

# Machine Learning (simplified - no Lightning)
torch==2.0.1
torchvision==0.15.2
torchaudio==0.0.2
scikit-learn==1.3.0
joblib==1.3.2

# Power System
pypsa==0.24.0

# Optional (for full system)
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
aiohttp==3.8.5
redis==4.6.0
psycopg2-binary==2.9.7